refactor: centralize normalization helpers
