+++ tests/test_data_pipeline.py
@@ -3,7 +3,7 @@ import pytest
 pytest.importorskip("pandas")
 import pandas as pd
 
-from backend.infrastructure.glpi import normalization as transform
+from backend.adapters import normalization as transform
 from backend.infrastructure.glpi.normalization import process_raw
